Header file main.h

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	int num, h, i, j;
+	unsigned long num, h, i, j;
 
 	h = 0;
 	i = 1;
